Adds the J2ObjCIncompatible annotation and support for stripping elements marked with it. 	Change on 2015/05/01 by kstanger <kstanger@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=92566422

+++ translator/src/main/java/com/google/devtools/j2objc/util/FileUtil.java
@@ -155,7 +155,7 @@ public class FileUtil {
    */
   public static void deleteTempDir(File dir) {
     // TODO(cpovirk): try Directories.deleteRecursively if a c.g.c.unix dep is OK
-    if (dir.exists()) {
+    if (dir != null && dir.exists()) {
       for (File f : dir.listFiles()) {
         if (f.isDirectory()) {
           deleteTempDir(f);
